requirements.txt
pdf2image
opencv-python
numpy
Pillow
